Break generic pieces of EstimateZernikesFamTask.py into EstimateZernikesBase.py
